Add Disconnect Screen
Add disconnect screen on world / server quit
Add config option to toggle
Convert ConfirmExitScreen to abstract ConfirmScreen with yesButtonCallback to decide what happens when the "yes" button is clicked
Add german translation
