+++ common/src/main/resources/assets/exit_confirm/lang/en_us.json
@@ -9,5 +9,8 @@
   "exit_confirm.config.screen.forceDirtBackground": "Force Dirt Background in Close Prompt",
   "exit_confirm.config.screen.forceTransparentBackground": "Force Transparent Background in Close Prompt",
   "exit_confirm.config.screen.allowNarrator": "Allow Narrator in Close Prompt",
-  "exit_confirm.config.screen.confirmDelay": "Close Confirmation Delay in Ticks"
+  "exit_confirm.config.screen.confirmDelay": "Close Confirmation Delay in Ticks",
+  "exit_confirm.config.prompt.disconnect": "Disconnect Confirmation",
+  "screen.disconnect_confirm.title": "Disconnect Confirmation",
+  "screen.disconnect_confirm.description": "Are you sure you want to disconnect?"
 }
