+++ common/src/main/resources/exit-confirmation.mixins.json
@@ -7,6 +7,7 @@
   "mixins": [
   ],
   "client": [
+    "PauseScreenMixin",
     "TitleScreenMixin"
   ],
   "injectors": {
